+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New method `Toolchain::uninstall` to remove a previously installed toolchain.
 - New method `Workspace::installed_toolchains` to get a list of all the
   toolchains in the workspace.
+- New error `PrepareError::PrivateGitRepository` when `Crate::fetch` is called
+  on a private or missing git repository.
 
 ### Changed